pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "revitpy"
dynamic = ["version"]
description = "Modern Python Framework for Revit Development"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "RevitPy Team" },
]
keywords = [
    "revit",
    "bim",
    "architecture",
    "construction",
    "automation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "asyncio-mqtt>=0.11.0",
    "aiofiles>=23.0.0",
    "loguru>=0.7.0",
    "httpx>=0.24.0",
    "websockets>=11.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "ruff>=0.0.260",
    "pre-commit>=3.0.0",
    "psutil>=5.9.0",
]
testing = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "hypothesis>=6.0.0",
    "factory-boy>=3.2.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
    "mkdocs-autorefs>=0.4.0",
]
ifc = ["ifcopenshell>=0.7.0"]
interop = ["specklepy>=2.0.0"]
sustainability = []  # uses existing httpx
cloud = []  # uses existing httpx, aiofiles
ai = []  # uses existing websockets, jinja2
all = ["ifcopenshell>=0.7.0", "specklepy>=2.0.0"]

[project.scripts]
revitpy = "revitpy.cli:main"

[project.urls]
Homepage = "https://github.com/revitpy/revitpy"
Documentation = "https://revitpy.readthedocs.io/"
Repository = "https://github.com/revitpy/revitpy"
Issues = "https://github.com/revitpy/revitpy/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["revitpy"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_no_return = true
strict_equality = true
show_error_codes = true
# Relax rules with many pre-existing violations; tighten incrementally
disable_error_code = [
    "attr-defined",
    "assignment",
    "var-annotated",
    "valid-type",
    "operator",
    "arg-type",
    "misc",
    "union-attr",
    "override",
    "return-value",
    "index",
    "call-overload",
    "no-redef",
    "exit-return",
    "has-type",
    "dict-item",
    "truthy-function",
    "type-var",
    "type-abstract",
    "import-untyped",
]

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["cli/templates/"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # flake8-bandit security rules
]
ignore = [
    "E501",  # line too long, handled by ruff format
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B904",  # raise from within except - TODO: fix these
    "E722",  # bare except - TODO: fix these
    "E741",  # ambiguous variable name
    "UP007", # Use X | Y for type annotations
    "B007",  # unused loop variable
    "B023",  # function not binding loop variable
    "B017",  # pytest.raises(Exception)
    "F811",  # redefinition of unused
    "S110",  # try-except-pass
    "S112",  # try-except-continue
    "S603",  # subprocess call
    "S607",  # partial executable path
    "N802",  # function name should be lowercase
    "N805",  # first argument should be self
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["S101", "D", "F401", "E402"]
"**/tests/**/*" = ["S101", "S105", "S106", "D", "F401", "E402"]
"bridge/**/*" = ["F401", "E402"]
"revitpy/performance/**/*" = ["F401"]
"cli/**/*" = ["F401", "E402", "B904"]
"revitpy-package-manager/**/*" = ["F401", "E402", "F403", "F405", "S"]
"proof-of-concepts/**/*" = ["S311"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["revitpy"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
